feat: Add react-chartjs-2 dependency and create Timer page
This commit adds the "react-chartjs-2" dependency to the project's package.json file. It also introduces a new Timer page, which includes a timer component that displays the number of seconds elapsed and a line chart showing the mortality rate over time. The Timer page uses the "react-chartjs-2" library to render the line chart.